Add old commits that got removed from master
